move assess and redline operations to assessment crud
