+++ app/Http/Controllers/Admin/Traits/UsesSaveAndNextAction.php
@@ -7,22 +7,24 @@ use Illuminate\Support\Str;
 trait UsesSaveAndNextAction
 {
 
-//    public function saveTabs()
-//    {
-//        $this->crud->set('tabs')
-//    }
+    // helper function - written so the removeAllSaveActions is more inline with the others below
+    public function removeAllSaveActions(): void
+    {
+        $this->crud->removeAllSaveActions();
+    }
+
 
     // You MUST call this AFTER creating all the tabbed fields
-    public function setupCustomSaveActions($operationRoute = 'edit')
+    public function addSaveAndNextAction($operationRoute = 'edit', $finalRedirect = null): void
     {
 
         $tabs = $this->crud->getTabs();
         $this->crud->set('tabs', $tabs);
         $this->crud->set('operationRoute', $operationRoute);
 
-        $this->crud->replaceSaveActions([
+        $this->crud->addSaveAction([
             'name' => 'save_and_next',
-            'redirect' => function ($crud, $request, $itemId){
+            'redirect' => function ($crud, $request, $itemId) use ($finalRedirect) {
                 $tabs = $this->crud->get('tabs');
                 $lastTab = $tabs[count($tabs) - 1];
 
@@ -44,11 +46,33 @@ trait UsesSaveAndNextAction
                         return $crud->route . "/" . $itemId . "/". $route ."#" . $next_tabs[$request->_current_tab];
                     }
                 }
-                return $crud->route ;
+                return $finalRedirect ?? $crud->route ;
             }, // what's the redirect URL, where the user will be taken after saving?
 
             // OPTIONAL:
             'button_text' => 'Save and Next', // override text appearing on the button
+            // You can also provide translatable texts, for example:
+            // 'button_text' => trans('backpack::crud.save_action_one'),
+            'visible' => function ($crud) {
+                return true;
+            }, // customize when this save action is visible for the current operation
+            'referrer_url' => function ($crud, $request, $itemId) {
+                return $crud->route;
+            }, // override http_referrer_url
+            'order' => 1, // change the order save actions are in
+        ]);
+    }
+
+    public function addSaveAndReturnToProjectListAction(): void
+    {
+        $this->crud->addSaveAction([
+            'name' => 'save_and_return_to_projects',
+            'redirect' => function ($crud, $request, $itemId){
+                return backpack_url('project');
+            }, // what's the redirect URL, where the user will be taken after saving?
+
+            // OPTIONAL:
+            'button_text' => 'Save and Return', // override text appearing on the button
             // You can also provide translatable texts, for example:
             // 'button_text' => trans('backpack::crud.save_action_one'),
             'visible' => function ($crud) {
